+++ src/test/java/test/org/fugerit/java/github/issue/export/TestGithubIssueGUI.java
@@ -56,6 +56,7 @@ public class TestGithubIssueGUI extends BasicTest {
 		    	.until( file::exists );
 			gui.dispose();
 			log.info( "file : {} -> {}", file.getCanonicalPath(), file.length() );
+			file.delete();
 		} );
 		
 	}